complete deleteCommentById

+++ backend/controllers/comments.js
@@ -2,9 +2,10 @@ const commentsModel = require("../models/commentSchema");
 
 //createComment
 const createComment = (req, res) => {
-  const { comment, product, createdBy } = req.body;
+  const { comment, product } = req.body;
   const myDate = new Date();
   const createdDate = myDate;
+  const createdBy = req.token.userId
   const newComment = new commentsModel({
     comment,
     product, //product ID
@@ -32,13 +33,23 @@ const createComment = (req, res) => {
 //deleteCommentById
 const deleteCommentById = (req, res) => {
   const { id } = req.params;
+  const createdBy = req.token.userId
   commentsModel
-    .findOneAndDelete({ _id: id })
-    .then(() => {
-      res.status(200).json({
-        success: true,
-        message: "comment deleted",
-      });
+    .findOneAndDelete({ _id: id , createdBy})
+    .then((result) => {
+        //if there is no result than mean the userId in createdBy is not the user who create this comment
+      if(result){
+        res.status(200).json({
+            success: true,
+            message: "comment deleted",
+          });
+      }
+      else{
+        res.status(403).json({
+            success: false,
+            message: "cant delete",
+          });
+      }
     })
     .catch((err) => {
       res.status(500).json({
@@ -51,4 +62,5 @@ const deleteCommentById = (req, res) => {
 
 module.exports = {
   createComment,
+  deleteCommentById
 };
